+++ src/courseware/sequence-navigation/CompleteIcon.jsx
