+++ src/courseware/sequence-navigation/SequenceNavigation.jsx
