+++ src/courseware/sequence-navigation/SequenceNavigationDropdown.jsx
