+++ src/courseware/sequence-navigation/SequenceNavigationTabs.jsx
