+++ src/courseware/sequence-navigation/UnitIcon.jsx
