+++ src/courseware/sequence-navigation/index.js
@@ -1,0 +1,2 @@
+export { default as SequenceNavigation } from './SequenceNavigation';
+export { default as UnitNavigation } from './UnitNavigation';
