+++ src/data/courseware/index.js
@@ -3,8 +3,6 @@ export {
   fetchSequence,
   checkBlockCompletion,
   saveSequencePosition,
-  addBookmark,
-  removeBookmark,
 } from './thunks';
 
 export { reducer } from './slice';
