+++ src/data/courseware/thunks.js
@@ -5,8 +5,6 @@ import {
   getSequenceMetadata,
   getBlockCompletion,
   updateSequencePosition,
-  createBookmark,
-  deleteBookmark,
 } from './api';
 import {
   addModelsMap, updateModel, updateModels,
@@ -134,77 +132,6 @@ export function saveSequencePosition(courseUsageKey, sequenceId, position) {
         model: {
           id: sequenceId,
           position: initialPosition,
-        },
-      }));
-    }
-  };
-}
-
-export function addBookmark(unitId) {
-  return async (dispatch) => {
-    // Optimistically update the bookmarked flag.
-    dispatch(updateModel({
-      modelType: 'units',
-      model: {
-        id: unitId,
-        bookmarked: true,
-        bookmarkedUpdateState: 'loading',
-      },
-    }));
-
-    try {
-      await createBookmark(unitId);
-      dispatch(updateModel({
-        modelType: 'units',
-        model: {
-          id: unitId,
-          bookmarked: true,
-          bookmarkedUpdateState: 'loaded',
-        },
-      }));
-    } catch (error) {
-      logError(error);
-      dispatch(updateModel({
-        modelType: 'units',
-        model: {
-          id: unitId,
-          bookmarked: false,
-          bookmarkedUpdateState: 'failed',
-        },
-      }));
-    }
-  };
-}
-
-export function removeBookmark(unitId) {
-  return async (dispatch) => {
-    // Optimistically update the bookmarked flag.
-    dispatch(updateModel({
-      modelType: 'units',
-      model: {
-        id: unitId,
-        bookmarked: false,
-        bookmarkedUpdateState: 'loading',
-      },
-    }));
-    try {
-      await deleteBookmark(unitId);
-      dispatch(updateModel({
-        modelType: 'units',
-        model: {
-          id: unitId,
-          bookmarked: false,
-          bookmarkedUpdateState: 'loaded',
-        },
-      }));
-    } catch (error) {
-      logError(error);
-      dispatch(updateModel({
-        modelType: 'units',
-        model: {
-          id: unitId,
-          bookmarked: true,
-          bookmarkedUpdateState: 'failed',
         },
       }));
     }
